Day 10 : Group Anagrams
Day 10 : Group Anagrams

You are given an array of strings strs[]. Your task is to group all the strings that are anagrams of each other. An anagram is a word or phrase formed by rearranging the letters of a different word or phrase, typically using all the original letters exactly once.
The goal is to return the grouped anagrams as a list of lists, where each sublist contains words that are anagrams of each other.

Input:
An array of strings strs[] consisting of lowercase English letters.

Output:
A list of lists, where each sublist contains strings that are anagrams of each other. The order of the output groups does not matter.

Examples:
Example 1
Input: strs[] = ["eat", "tea", "tan", "ate", "nat", "bat"]
Output: [["eat", "tea", "ate"], ["tan", "nat"], ["bat"]]